moved findLast util
